Increase the lidar data DMA handling space, and modify the MSOP test code

+++ driver/lib/libsfconfig.cpp
@@ -216,7 +216,8 @@ int Lidar_Get_DIFOP(int fd, RS16DifopPkt *difop)
 int Lidar_Set_fps(int fd, unsigned int fps)
 {
     int ret = 0;
-    unsigned int rpm;
+    //TODO
+    unsigned short rpm;
     switch (fps)
     {
     case 20:
